requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
playwright==1.46.0
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.8.2
pyyaml==6.0.2
